use run in contract
