+++ deal/linter/_extractors/imports.py
@@ -10,6 +10,8 @@ def get_imports(body: list) -> Iterator[Token]:
     for expr in traverse(body):
         token_info = dict(line=expr.lineno, col=expr.col_offset)
         if isinstance(expr, astroid.ImportFrom):
-            yield Token(value=expr.modname, **token_info)
+            dots = '.' * (expr.level or 0)
+            yield Token(value=dots + expr.modname, **token_info)
         if isinstance(expr, ast.ImportFrom):
-            yield Token(value=expr.module, **token_info)
+            dots = '.' * expr.level
+            yield Token(value=dots + expr.module, **token_info)
